+++ src/main/java/insane96mcp/iguanatweaksexpanded/mixin/ProjectileMixin.java
@@ -1,0 +1,25 @@
+package insane96mcp.iguanatweaksexpanded.mixin;
+
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.BurstOfArrows;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.projectile.Projectile;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.HitResult;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(Projectile.class)
+public abstract class ProjectileMixin extends Entity {
+
+    public ProjectileMixin(EntityType<?> pEntityType, Level pLevel) {
+        super(pEntityType, pLevel);
+    }
+
+    @Inject(method = "onHit", at = @At(value = "HEAD"))
+    public void onOnHit(HitResult pResult, CallbackInfo ci) {
+        BurstOfArrows.trySummonArrows((Projectile) (Object) this);
+    }
+}